+++ src/main/java/com/company/firstproject/service/implementation/BasicDerivativeService.java
@@ -4,10 +4,10 @@ import com.company.firstproject.derivative.Derivative;
 import com.company.firstproject.entity.obligations.Obligation;
 import com.company.firstproject.service.DerivativeService;
 
+import java.lang.reflect.Array;
 import java.util.List;
 
 public class BasicDerivativeService implements DerivativeService {
-
     private Derivative derivative;
 
     public BasicDerivativeService(Derivative derivative) {
@@ -15,7 +15,7 @@ public class BasicDerivativeService implements DerivativeService {
     }
 
     @Override
-    public List<Obligation> findByRiskRange(float startValue, float endValue) {
+    public List<Obligation> findByRiskRange(double startValue, double endValue) {
         return derivative.findByRiskRange(startValue, endValue);
     }
 
@@ -38,6 +38,4 @@ public class BasicDerivativeService implements DerivativeService {
     public String toString() {
         return derivative.toString();
     }
-
-
 }
